[myListSlice] Use one array for movies and tvs
The myList should show the items in the order the user add items to it,
I was using two seperate arrays to store the movies and tvs items, instead i'm using one array for both and adding the parent to the object to know the type of the item.
In that way the order will be correct and i will not have to store any other informations to know the order of items or fo any kind of sorting.