[redux] Store myList on the device and use thunk
I wasn't stroing the myList on the device, i was just add items to the array then show them at MyListScreen, so when the app is closed and re-opened the MyListScreen has nothing to show.
So, it was obviouse that i need to store the data on local storage, and i'm using AsyncStorage to store the data and thunk as a middleware to make the async logic in the code.

P.S : I need to review this commit before continue the work, i added some comment on the code, check them.