[Search] Change The way search works
Before : Catch the text on `SearchBar`, update the query and get the search results and save the results in the store, then in `SearchList` get the query and the results from the store, and on scrolling down, fetch the next page

Now : The `SearchBar` is only responsible for updating the query, no additional operations there. In `SearchList` get the query from the store and fetch the search results whenever the query change.

This way we only stor `query` instead of `query`, `searchResults`, and `totalPages` to just using them in `SearchList`.

OTOH : The name of `searchResultsSlice` is not accurate, fixing that in the next commit.